+++ instapy/like_util.py
@@ -197,6 +197,7 @@ def get_links_for_tag(browser, tag, amount, media=None, skip_top_posts=True):
         body_elem.send_keys(Keys.END)
         sleep(2)
         load_button.click()
+
         # update server calls
         update_activity()
 
@@ -400,8 +401,11 @@ def check_link(browser,
     sleep(2)
 
     """Check if the Post is Valid/Exists"""
-    post_page = browser.execute_script(
-        "return window._sharedData.entry_data.PostPage")
+    try:
+        post_page = browser.execute_script(
+        	"return window._sharedData.entry_data.PostPage")
+    except:
+        post_page = None
     if post_page is None:
         print('Unavailable Page: {}'.format(link.encode('utf-8')))
         return True, None, None, 'Unavailable Page'
@@ -515,13 +519,12 @@ def check_link(browser,
 
     return False, user_name, is_video, 'None'
 
-
 def like_image(browser, username, blacklist):
     """Likes the browser opened image"""
     like_elem = browser.find_elements_by_xpath(
         "//a[@role='button']/span[text()='Like']/..")
-    liked_elem = browser.find_elements_by_xpath(
-        "//a[@role='button']/span[text()='Unlike']")
+    #liked_elem = browser.find_elements_by_xpath(
+    #    "//a[@role='button']/span[text()='Unlike']")
 
     if len(like_elem) == 1:
         like_elem[0].send_keys("\n")
@@ -534,9 +537,9 @@ def like_image(browser, username, blacklist):
             )
         sleep(2)
         return True
-    elif len(liked_elem) == 1:
-        print('--> Already Liked!')
-        return False
+    #elif len(liked_elem) == 1:
+    #    print('--> Already Liked!')
+    #    return False
     else:
         print('--> Invalid Like Element!')
         return False
